+++ term3d.h
@@ -31,7 +31,7 @@ char		*read_file(const char *file_name);
 size_t		count_new_line(char *content);
 void		exit_with_error(const char *massage);
 int			convert_to_mapindex(const t_vector *v);
-t_vector	affine4(const double matrix[], const t_vector *v);
+t_vector	affine(const double matrix[], const t_vector *v);
 void		init_matrix(double matrix[]);
 void		matrix_rotate_y(double matrix[]);
 void		matrix_translate(double matrix[], const double x, const double y, const double z);
